Feat: 마이리스트, 콜라보리스트 페이지로 각각 분리
