+++ src/app/[userNickname]/_components/Card.tsx
