+++ src/app/[userNickname]/_components/CardItem.css.ts
