+++ src/app/[userNickname]/_components/CardItem.tsx
