+++ src/app/[userNickname]/_components/Categories.tsx
