+++ src/app/[userNickname]/_components/Profile.tsx
