+++ src/app/[userNickname]/mockData/categories.ts
