+++ src/app/[userNickname]/mockData/lists.ts
