+++ src/app/[userNickname]/mockData/mockDataTypes.ts
