+++ src/app/[userNickname]/mockData/user.ts
